교수 수정

+++ src/main/java/com/green/smartGrade/professor/ProfessorService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.List;
 
 @Service
@@ -37,50 +38,69 @@ public class ProfessorService {
     }
 
 
-        public ProfessorUpRes upProfessor(ProfessorParam param) {
+//        public ProfessorUpRes upProfessor(ProfessorParam param) {
+//        ProfessorUpDto dto = new ProfessorUpDto();
+//        dto.setPhone(param.getPhone());
+//        dto.setEmail(param.getEmail());
+//        dto.setAddress(param.getAddress());
+//        dto.setIprofessor(param.getIprofessor());
+//
+//        int result = mapper.upProfessor(dto);
+//        if (result == 1 ) {
+//            ProfessorUpRes res = new ProfessorUpRes(dto);
+//            return res;
+//        }
+//        return null;
+//    }
+    public ProfessorUpRes upProfessor(MultipartFile pic, ProfessorParam param) {
+
+        int result = 0;
         ProfessorUpDto dto = new ProfessorUpDto();
+        dto.setAddress(param.getAddress());
         dto.setPhone(param.getPhone());
         dto.setEmail(param.getEmail());
-        dto.setAddress(param.getAddress());
         dto.setIprofessor(param.getIprofessor());
+        result = mapper.upProfessor(dto);
 
-        int result = mapper.upProfessor(dto);
         if (result == 1 ) {
+            if (pic != null) {
+
+                String centerPath = String.format("professor/%d", param.getIprofessor());
+                String dicPath = String.format("%s/%s", FileUtils.getAbsolutePath(fileDir), centerPath);
+                String temp = "0";
+
+                File dic = new File(dicPath);
+                if(!dic.exists()) {
+                    dic.mkdirs();
+                }
+
+                String originFileName = pic.getOriginalFilename();
+                String savedFileName = FileUtils.makeRandomFileNm(originFileName);
+                String savedFilePath = String.format("%s/%s", centerPath, savedFileName);
+                String targetPath = String.format("%s/%s", FileUtils.getAbsolutePath(fileDir), savedFilePath);
+
+                File target = new File(targetPath);
+
+                try {
+                    pic.transferTo(target);
+                } catch (IOException e) {
+                    throw new RuntimeException(temp);
+                }
+                dto.setPic(savedFileName);
+
+                try {
+                    result = mapper.upProfessor(dto);
+                    if (result == Integer.parseInt(temp)){
+                        throw new Exception("스티커 사진 불가");
+                    }
+                } catch (Exception e) {
+                    target.delete();
+                }
+            }
             ProfessorUpRes res = new ProfessorUpRes(dto);
             return res;
         }
         return null;
-    }
-    public String upPicProfessor(MultipartFile pic , ProfessorPicDto dto) {
-        String centerPath = String.format("professor/%d", dto.getIprofessor());
-        String dicPath = String.format("%s/%s", FileUtils.getAbsolutePath(fileDir), centerPath);
-        String temp = "0";
-        File dic = new File(dicPath);
-        if(!dic.exists()) {
-            dic.mkdirs();
-        }
-
-        String originFileName = pic.getOriginalFilename();
-        String savedFileName = FileUtils.makeRandomFileNm(originFileName);
-        String savedFilePath = String.format("%s/%s", centerPath, savedFileName);
-        String targetPath = String.format("%s/%s", FileUtils.getAbsolutePath(fileDir), savedFilePath);
-        File target = new File(targetPath);
-        try {
-            pic.transferTo(target);
-        } catch (Exception e) {
-            return temp;
-        }dto.setPic(savedFilePath);
-        try {
-            int result = mapper.upPicProfessor(dto);
-            if (result == Integer.parseInt(temp)) {
-                throw new Exception("스티커 사진 등록 불가");
-            }
-        } catch (Exception e) {
-            target.delete();
-            return temp;
-        }
-        return savedFilePath;
-
     }
     public professorSelRes selAllProfessor(ProfessorSelDto dto,int page) {
         int maxPage = mapper.ProfessorCount();
